goes through calculation of alpha_i
most of this has been cleaning stuff up and making sure I understand
things. the big change is the addition of photoreceptor lattice. as part
of that I split up the absorb function into two: the first calculates
the retinal image (just convolves the luminance image and the
pointspread function), while the second scales this by the variety of
constants in equation 2 and then samples the places where there are
photoreceptors