adds start of optimization, retinal image approx
this commit adds two main things: the start of the optimization we'll
use to find corresponding inputs given a task and one of the inputs (as
far as we can tell, there's no closed form solution) and the retinal
image approximation. the approximation does not use the convolution,
since then the optimization would take way too long. also includes some
notes on what to do next.