Gets resolution task working
we figured out was wrong with our resolution task and it basically came
down to how we were making the shapes of the pointspread function and
the receptor lattice lined up. we were incorrectly embedding the
pointspread function in a sea of zeros, which is why we had so many
receptors that had zero absorption. changing it so the pointspread
function was the right shape from the beginning made it possible to fix
the issue, but meant we had to spend some more time in our retinal_image
function dealing with the case when the shapes were identical.

also had to add a special case for when the log_ratio of alpha and beta
is all zeros (in which case dprime is 0)

it all works now, but we'll have to spend a bit more time working
through how to make figure 5